refactor/#17: 유저 검증 로직을 AuthService에서 UserService로 이동
